Inntekt kan ha desimaler, reglene håndterer dette i sine beregninger, men Arena forventer heltall i inntektsperiodene.
Resolves navikt/dagpenger#154

+++ src/main/kotlin/no/nav/dagpenger/regel/api/internal/models/Inntekt.kt
@@ -1,7 +1,9 @@
 package no.nav.dagpenger.regel.api.internal.models
 
+import java.math.BigDecimal
+
 data class Inntekt(
-    val inntekt: Int,
+    val inntekt: BigDecimal,
     val periode: Int, // todo: enum?
     val inntektsPeriode: InntektsPeriode,
     val inneholderFangstOgFisk: Boolean,
