+++ src/test/kotlin/no/nav/dagpenger/regel/api/internal/ExtractMinsteinntektOgPeriodeTest.kt
@@ -119,19 +119,19 @@ class ExtractMinsteinntektOgPeriodeTest {
             oppfyllerMinsteinntekt = true,
             minsteinntektInntektsPerioder = listOf(
                 Inntekt(
-                    inntekt = 600000,
+                    inntekt = 600000.toBigDecimal(),
                     periode = 1,
                     inntektsPeriode = InntektsPeriode(YearMonth.of(2018, 5), YearMonth.of(2019, 5)),
                     inneholderFangstOgFisk = true
                 ),
                 Inntekt(
-                    inntekt = 500000,
+                    inntekt = 500000.toBigDecimal(),
                     periode = 2,
                     inntektsPeriode = InntektsPeriode(YearMonth.of(2017, 5), YearMonth.of(2018, 5)),
                     inneholderFangstOgFisk = false
                 ),
                 Inntekt(
-                    inntekt = 400000,
+                    inntekt = 400000.toBigDecimal(),
                     periode = 3,
                     inntektsPeriode = InntektsPeriode(YearMonth.of(2016, 5), YearMonth.of(2017, 5)),
                     inneholderFangstOgFisk = true
@@ -165,19 +165,19 @@ class ExtractMinsteinntektOgPeriodeTest {
             oppfyllerMinsteinntekt = false,
             minsteinntektInntektsPerioder = listOf(
                 Inntekt(
-                    inntekt = 0,
+                    inntekt = 0.toBigDecimal(),
                     periode = 1,
                     inntektsPeriode = InntektsPeriode(YearMonth.of(2018, 5), YearMonth.of(2019, 5)),
                     inneholderFangstOgFisk = true
                 ),
                 Inntekt(
-                    inntekt = 0,
+                    inntekt = 0.toBigDecimal(),
                     periode = 2,
                     inntektsPeriode = InntektsPeriode(YearMonth.of(2017, 5), YearMonth.of(2018, 5)),
                     inneholderFangstOgFisk = false
                 ),
                 Inntekt(
-                    inntekt = 0,
+                    inntekt = 0.toBigDecimal(),
                     periode = 3,
                     inntektsPeriode = InntektsPeriode(YearMonth.of(2016, 5), YearMonth.of(2017, 5)),
                     inneholderFangstOgFisk = true
